+++ Hamburgerci.Domain/Entities/Concrete/EkstraMalzeme.cs
@@ -7,7 +7,7 @@ namespace Hamburgerci.Entities.Concrete
     {
         public int Id { get; set; }
         public string Adi { get; set; }
-        public decimal Fiyati { get; set; }
+        public double Fiyati { get; set; }
         public ParaBirimi ParaBirimi { get; set; }
 
         public ICollection<Siparis> Siparisler { get; set; }
